Fix maven repo ordering and crash for initial file cleanup routine

+++ app/src/main/java/com/androidexperiments/tunnelvision/utils/AndroidUtils.java
@@ -53,10 +53,11 @@ public class AndroidUtils
                 + File.separator;
         File directory = new File(path);
         File[] files = directory.listFiles();
-
-        for (int i = 0; i < files.length; i++) {
-            if (files[i].length() == 0) {
-                files[i].delete();
+        if (files != null) {
+            for (int i = 0; i < files.length; i++) {
+                if (files[i].length() == 0) {
+                    files[i].delete();
+                }
             }
         }
